优化代码

+++ net/bpf/vm_instructions.go
@@ -96,7 +96,7 @@ func jumpIfCommon(cond JumpTest, skipTrue, skipFalse uint8, regA uint32, value u
 
 func loadAbsolute(ins LoadAbsolute, in []byte) (uint32, bool) {
 	offset := int(ins.Off)
-	size := int(ins.Size)
+	size := ins.Size
 
 	return loadCommon(in, offset, size)
 }
@@ -123,7 +123,7 @@ func loadExtension(ins LoadExtension, in []byte) uint32 {
 
 func loadIndirect(ins LoadIndirect, in []byte, regX uint32) (uint32, bool) {
 	offset := int(ins.Off) + int(regX)
-	size := int(ins.Size)
+	size := ins.Size
 
 	return loadCommon(in, offset, size)
 }
@@ -155,7 +155,7 @@ func loadCommon(in []byte, offset int, size int) (uint32, bool) {
 	case 2:
 		return uint32(binary.BigEndian.Uint16(in[offset : offset+size])), true
 	case 4:
-		return uint32(binary.BigEndian.Uint32(in[offset : offset+size])), true
+		return binary.BigEndian.Uint32(in[offset : offset+size]), true
 	default:
 		panic(fmt.Sprintf("invalid load size: %d", size))
 	}
